Sound slider fixed

+++ Zephyr2/source/AudioEngine.h
@@ -56,7 +56,7 @@ public:
 	int PlayAudio(const string &strSoundName, const Vector3 &vPosition = Vector3{ 0, 0, 0 }, float fVolumedB = 5.0f);
 	void SetChannel3dPosition(int nChannelId, const Vector3 &vPosition);
 	void SetChannelvolume(int nChannelId, float fVolumedB);
-	void SetSoundVolume(int volume);
+	void SetSoundVolume(float volume);
 	float dbToVolume(float db);
 	float VolumeTodb(float volume);
 	FMOD_VECTOR VectorToFmod(const Vector3 &vPosition);
